07/04/19 16:45

+++ Personal_Excersises/primes.cpp
@@ -1,21 +1,21 @@
 #include <iostream>
 #include <vector>
 #include <cmath>
-
+//Calculate the sum of all primes up to 2 million
 int main()
 {
-   uint64_t answ=0;
-   const uint64_t n = 2000000;
+   size_t answ=0;
+   const size_t n = 2000000;
    std::vector<bool> vec(n+1,true);
    vec[0]=false;
    vec[1]=false;
   
-   for(uint64_t i =2;i<=std::floor(std::sqrt(n));i++)  
+   for(size_t i =2;i<=std::floor(std::sqrt(n));i++)
    {       
        if(vec[i]==true)
        {           
            int m = 0;
-           for(uint64_t j=std::pow(i,2);j<=n;j=std::pow(i,2)+(m*i))
+           for(size_t j=std::pow(i,2);j<=n;j=std::pow(i,2)+(m*i))
            {
                vec[j]=false;
                m++;
@@ -23,7 +23,7 @@ int main()
        }
    }
    ///////////////////////////////////////////////////////////////////////////
-   for(uint64_t i =0;i<vec.size();i++)
+   for(size_t i =0;i<vec.size();i++)
    {
        if(vec[i]==true)
        {
